Move properly to resolverHandler
