+++ src/main/java/com/minecolonies/coremod/colony/requestsystem/management/handlers/RequestHandler.java
@@ -33,11 +33,6 @@ public class RequestHandler implements IRequestHandler
 
     private final IStandardRequestManager manager;
 
-    /**
-     * Additional temporary blacklist of the request handler.
-     */
-    private List<IToken<?>> tempBlackList = new ArrayList<>();
-
     public RequestHandler(final IStandardRequestManager manager) {this.manager = manager;}
 
     @Override
@@ -153,7 +148,7 @@ public class RequestHandler implements IRequestHandler
         for (@SuppressWarnings(RAWTYPES) final IRequestResolver resolver : resolvers)
         {
             //Skip when the resolver is in the blacklist.
-            if (resolverTokenBlackList.contains(resolver.getId()) || tempBlackList.contains(resolver.getId()))
+            if (resolverTokenBlackList.contains(resolver.getId()) || manager.getResolverHandler().isBeingRemoved(resolver.getId()))
             {
                 continue;
             }
@@ -530,18 +525,6 @@ public class RequestHandler implements IRequestHandler
                 onRequestCancelled(req.getId());
             }
         }
-    }
-
-    @Override
-    public void addToTempBlackList(final IToken<?> resolverToken)
-    {
-        tempBlackList.add(resolverToken);
-    }
-
-    @Override
-    public void removeFromTempBlackList(final IToken<?> resolverToken)
-    {
-        tempBlackList.remove(resolverToken);
     }
 
     /**
